Begin modularizing providers
