+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod handlers;
+pub mod providers;
 
 extern crate emailaddress;
 extern crate hyper;
@@ -15,25 +16,17 @@ extern crate url;
 extern crate urlencoded;
 
 use emailaddress::EmailAddress;
-use hyper::client::Client as HttpClient;
-use iron::modifiers;
-use iron::prelude::{IronResult, Response};
-use iron::status;
-use iron::Url;
-use lettre::transport::EmailTransport;
 use openssl::crypto::hash;
 use openssl::crypto::pkey::PKey;
 use serde_json::de::from_reader;
 use serde_json::value::Value;
 use rand::{OsRng, Rng};
-use redis::{Client, Commands};
+use redis::Client;
 use rustc_serialize::base64::{self, ToBase64};
 use std::collections::BTreeMap;
 use std::fs::File;
 use std::io::{BufReader, Write};
 use std::iter::Iterator;
-use url::percent_encoding::{utf8_percent_encode, QUERY_ENCODE_SET};
-use urlencoded::QueryMap;
 
 
 /// Configuration data for a "famous" identity provider.
@@ -124,7 +117,7 @@ impl AppConfig {
 /// Helper function to build a session ID for a login attempt.
 ///
 /// Put the email address, the client ID (RP origin) and some randomness into
-/// a SHA256 hash, and encode it with URL-safe bas64 encoding. This is used
+/// a SHA256 hash, and encode it with URL-safe base64 encoding. This is used
 /// as the key in Redis, as well as the state for OAuth authentication.
 fn session_id(email: &EmailAddress, client_id: &str) -> String {
     let mut rng = OsRng::new().unwrap();
@@ -137,81 +130,6 @@ fn session_id(email: &EmailAddress, client_id: &str) -> String {
     hasher.write(&rand_bytes).unwrap();
     hasher.finish().to_base64(base64::URL_SAFE)
 }
-
-
-/// Helper method to issue an OAuth authorization request.
-///
-/// When an authentication request comes in and matches one of the "famous"
-/// identity providers configured in the `AppConfig`, we redirect the client
-/// to an Authentication Request URL, which we discover by reading the
-/// provider's configured Discovery URL. We pass in the client ID we received
-/// when pre-registering for the provider, as well as a callback URL which the
-/// user will be redirected back to after confirming (or denying) the
-/// Authentication Request. Included in the request is a nonce which we can
-/// later use to definitively match the callback to this request.
-fn oauth_request(app: &AppConfig, params: &QueryMap) -> IronResult<Response> {
-    let email_addr = EmailAddress::new(&params.get("login_hint").unwrap()[0]).unwrap();
-    let client_id = &params.get("client_id").unwrap()[0];
-    let session = session_id(&email_addr, client_id);
-
-    // Store the nonce and the RP's `redirect_uri` in Redis for use in the
-    // callback handler.
-    let key = format!("session:{}", session);
-    let _: String = app.store.hset_multiple(key.clone(), &[
-        ("email", email_addr.to_string()),
-        ("client_id", client_id.clone()),
-        ("redirect", params.get("redirect_uri").unwrap()[0].clone()),
-    ]).unwrap();
-    let _: bool = app.store.expire(key.clone(), app.expire_keys).unwrap();
-
-    // Retrieve the provider's Discovery document and extract the
-    // `authorization_endpoint` from it.
-    // TODO: cache other data for use in the callback handler, so that we
-    // don't have to request the Discovery document twice. We could even store
-    // per-provider data in Redis so we can amortize the cost of discovery.
-    let provider = &app.providers[&email_addr.domain];
-    let client = HttpClient::new();
-    let rsp = client.get(&provider.discovery).send().unwrap();
-    let val: Value = from_reader(rsp).unwrap();
-    let config = val.as_object().unwrap();
-    let authz_base = config["authorization_endpoint"].as_string().unwrap();
-
-    // Create the URL to redirect to, properly escaping all parameters.
-    let auth_url = Url::parse(&vec![
-        authz_base,
-        "?",
-        "client_id=",
-        &utf8_percent_encode(&provider.client_id, QUERY_ENCODE_SET).to_string(),
-        "&response_type=code",
-        "&scope=",
-        &utf8_percent_encode("openid email", QUERY_ENCODE_SET).to_string(),
-        "&redirect_uri=",
-        &utf8_percent_encode(&format!("{}/callback", &app.base_url),
-                             QUERY_ENCODE_SET).to_string(),
-        "&state=",
-        &utf8_percent_encode(&session, QUERY_ENCODE_SET).to_string(),
-        "&login_hint=",
-        &utf8_percent_encode(&email_addr.to_string(), QUERY_ENCODE_SET).to_string(),
-    ].join("")).unwrap();
-    // Using 302 Found for redirection here. Note that, per RFC 7231, a user
-    // agent MAY change the request method from POST to GET for the subsequent
-    // request.
-    Ok(Response::with((status::Found, modifiers::Redirect(auth_url))))
-}
-
-
-/// Characters eligible for inclusion in the email loop one-time pad.
-///
-/// Currently includes all numbers, lower- and upper-case ASCII letters,
-/// except those that could potentially cause confusion when reading back.
-/// (That is, '1', '5', '8', '0', 'b', 'i', 'l', 'o', 's', 'u', 'B', 'D', 'I'
-/// and 'O'.)
-const CODE_CHARS: &'static [char] = &[
-    '2', '3', '4', '6', '7', '9', 'a', 'c', 'd', 'e', 'f', 'g', 'h', 'j', 'k',
-    'm', 'n', 'p', 'q', 'r', 't', 'v', 'w', 'x', 'y', 'z', 'A', 'C', 'E', 'F',
-    'G', 'H', 'J', 'K', 'L', 'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W',
-    'X', 'Y', 'Z',
-];
 
 
 /// Helper method to create a JWT from given header and payload.
